requirements.txt
google-adk==1.14.0
google-cloud-logging
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
google-cloud-storage
sqlalchemy
pg8000
google-cloud-alloydb-connector[pg8000]
gunicorn